@@ -3,27 +3,26 @@
   "private": true,
   "dependencies": {
     "@rails/webpacker": "5.2.1",
+    "@vue/compiler-sfc": "^3.0.5",
     "axios": "^0.21.1",
     "bootstrap": "^5.0.0-beta1",
     "css-loader": "^5.0.1",
     "kitsu-core": "^8.1.1",
-    "less": "^3.13.0",
-    "less-loader": "^7.1.0",
+    "node-sass": "^5.0.0",
     "qs": "^6.9.3",
-    "view-design": "^4.1.0",
-    "vue": "^2.6.1",
-    "vue-clipboard2": "^0.3.1",
+    "sass-loader": "^10.1.1",
+    "view3": "https://github.com/omohokcoj/view3",
+    "vue": "^3.0.0",
     "vue-content-placeholders": "^0.2.1",
-    "vue-loader": "^15.1.0",
-    "vue-router": "^3.4.9",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.6.0"
+    "vue-loader": "^16.1.2",
+    "vue-router": "^4.0.0",
+    "vuex": "^4.0.0-rc.2"
   },
   "scripts": {
     "eslint": "./node_modules/eslint/bin/eslint.js \"app/javascript/**/*.{js,vue}\" --fix",
-    "stylelint": "./node_modules/stylelint/bin/stylelint.js \"app/javascript/**/*.{scss,css,less}\" --fix"
+    "stylelint": "./node_modules/stylelint/bin/stylelint.js \"app/javascript/**/*.{scss,css}\" --fix"
   },
-  "version": "1.0.2",
+  "version": "1.0.3",
   "devDependencies": {
     "babel-eslint": "^10.1.0",
     "eslint": "^7.15.0",
