@@ -1,13 +1,15 @@
-import 'ui/index.less'
+import 'ui/misc/styles/bootstrap'
+import 'vue-content-placeholders/src/styles'
 import 'ui/index.scss'
 
 import app from 'ui/index'
+import store from 'ui/store'
 
 document.addEventListener('DOMContentLoaded', () => {
   if (document.getElementById('app')) {
-    app.$store.dispatch('loadCurrentUser')
-    app.$store.dispatch('loadWebsites')
+    store.dispatch('loadCurrentUser')
+    store.dispatch('loadWebsites')
 
-    app.$mount('#app')
+    app.mount('#app')
   }
 })
