@@ -1,6 +1,6 @@
 <template>
   <div class="layout">
-    <Header v-if="!['setup', 'shared'].includes($route.name)" />
+    <MainHeader v-if="!['setup', 'shared'].includes($route.name)" />
     <div class="container py-3 pb-5">
       <RouterView />
     </div>
@@ -8,12 +8,12 @@
 </template>
 
 <script>
-import Header from './navigation/components/header'
+import MainHeader from './navigation/components/header'
 
 export default {
   name: 'App',
   components: {
-    Header
+    MainHeader
   }
 }
 </script>
