@@ -33,7 +33,7 @@
 
 <script>
 import api from 'ui/api'
-import { Dialog, Message } from 'ui/misc/scripts/dynamic_components'
+import { Dialog, Message } from 'view3/src/plugins/dynamic-components'
 
 export default {
   name: 'User',
